Fixed a big rtp demuxer.
DEMUX_GET_PTS_DELAY should returns microsecond.

You should NOT remove caching to PTS to create PCR.
It was probably a workaround for the previous bug. It could explains
the problems with TS over RTP (for TS with small pcr/pts delay that is, not
created by VLC)

Implemented missing DEMUX_CAN_* query.