Qt Jpeg support on Windows.
This is a support to statically link jpeg plugin on windows to the plugin in order to make Qt read jpeg files (needed for album art).
This of course, will increase the binary size, but since dionoea is lazy and doesn't want to make easily conversion from jpeg to png|xpm|bmp, here we go. :D
Better solution should be found on master.