Change the default hotkeys to avoid conflicts.
1) Every Human Interface Guideline on Unix/Linux/Windows reserves alt key for menu and button accessibility... Try to NOT use those...
Moreover, you have NO idea how the translators are going to translate your accelerators...

On the other hand, with this long list of keys it was become more than difficult... I'll document on the wiki.

2) Use keys next-to-each-other like-a-cross for cropping

Those keys are making a cross in most layouts ( on the contrary to the previous ones).

3) At the same time, use easier snapshot key. (shift+s is easier than Ctrl+alt+s)

4) Comment previous changes and remove whitespace.

Nota Bene:
I have no idea on how are the default on the mac and if the are conflicts there too, so I won't make changes there.