contrib: add fontconfig cache and conf locations.
For Mac OS X, let's now use the default X11 locations for the fonts.conf and the cache files.
Those that do not have X11 installed (old default < 10.4) are simply out of luck.
They will only be able to use the files in /System/Library/Fonts and will need to rebuild
their cache every time.

We should also fix this for windows. I'm looking for other projects that use fontconfig on
Windows to see if I can get some hints on how to do this.