Win32 Vout: Try to fix the display of the MI on exit of movie in FS mode.
This is a big, a huge HACK around the windows Video Output,
Since the Vout fusion the Vout Windows, with the main windows
(very nice indeed for Alt-Tab), at the end of the film, when Vout quits
fullscreen, MI isn't properly put in the good state again...
(hidden still)...
Of course this is not good, but we will have to live with it temporarly...

Since we have no windows developer, too bad. And if you don't have the Fullscreen controller, and you don't have the systray icon, you are screwed up.

I will try a proper fix later.