RTSP authentication with DSS
This patch enables a working RTSP authentication with the Darwin
streaming server (DSS).

When the VLC is compiled against Live555 streaming media library
(live.2008.09.02), the authentication fails when both username and
password are asked via a VLC dialog as the DSS requires authorization
also for SETUP, PLAY, and TEARDOWN commands and the necessary
information isn't delivered correctly to the Live555 library. The VLC
used Live555 interface in such a way, that the authenticator was valid
only for the DESCRIBE when I debugged the library with vlc-0.9.2. The
problem still exists in vlc-0.9.5.

Signed-off-by: Rolf Ahrenberg <Rolf.Ahrenberg@saunalahti.fi>
Signed-off-by: Rémi Denis-Courmont <rdenis@simphalempin.com>