RTP out: improve soft-error handling
When the socket is congested, we should simply skip sending.
When there is a soft-error, we should resend and simply ignore.

Instead, we were dropping the socket if the failure remained on the
second try, which was always the case in the congestion scenario.
This should fix the socket dropping problem described at:
http://forum.videolan.org/viewtopic.php?f=4&t=56493
(cherry picked from commit 9fdc8915682d3d3beae69f6956616b660ec9963b)